Added alien movement

+++ Source/Constants.h
@@ -7,6 +7,6 @@
 
 static constexpr int WINDOW_WIDTH = 1024;  /**< Defines how wide the game window is. */
 static constexpr int WINDOW_HEIGHT = 768;  /**< Defines the height of the game window. */
-static constexpr float WINDOW_MARGIN = 20.0f;
-static constexpr float WINDOW_LEFT_BOUNDARY = WINDOW_MARGIN;
-static constexpr float WINDOW_RIGHT_BOUNDARY = WINDOW_WIDTH - WINDOW_LEFT_BOUNDARY;
+static constexpr int WINDOW_MARGIN = 20;
+static constexpr int WINDOW_LEFT_BOUNDARY = WINDOW_MARGIN;
+static constexpr int WINDOW_RIGHT_BOUNDARY = WINDOW_WIDTH - WINDOW_LEFT_BOUNDARY;
